requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
# Editable install with no version control (project_name==0.1.0)
-e /home/brian/code/redis_sql_python_flask_poc
psycopg2-binary==2.9.9
redis==5.0.7
requests==2.32.3
setuptools==72.0.0
urllib3==2.2.2
Werkzeug==3.0.3
